Cleanup batch processing code
